requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
PyJWT==1.7.1
typing-extensions==3.7.2
yarl==1.3.0
